modify businesses styles, lift states up
