+++ src/pages/Businesses/BusinessesList.jsx
@@ -1,8 +1,4 @@
-import React, { useEffect, useState } from 'react';
-import { connect } from 'react-redux';
-import { FormControlLabel, Switch } from '@material-ui/core';
-import { getBusinessesAct } from '../../state/businesses/businessActions';
-import isEqual from 'lodash.isequal';
+import React from 'react';
 import { TableComponent } from '../../components/Table/index';
 
 const headCells = [
@@ -13,58 +9,18 @@ const headCells = [
 ];
 
 const BusinessesList = props => {
-    const { businesses, isLoggedIn, getBusinessesAct } = props;
-    const [businessesList, setBusinessesList] = useState([]);
-    const [dense, setDense] = useState(false);
-
-    useEffect(() => {
-        getBusinessesAct();
-    }, [getBusinessesAct]);
-
-    useEffect(() => {
-        if (!isLoggedIn) {
-            const localBusinesses = JSON.parse(
-                window.localStorage.getItem('businesses'),
-            );
-            if (!isEqual(localBusinesses, businessesList)) {
-                setBusinessesList(localBusinesses);
-            }
-        } else {
-            if (!isEqual(businesses, businessesList)) {
-                setBusinessesList(businesses);
-            }
-        }
-    }, [isLoggedIn, businesses, businessesList, getBusinessesAct]);
-
-    const handleChangeDense = event => {
-        setDense(event.target.checked);
-    };
+    const { businessesList, selected, setSelected, dense } = props;
 
     return (
-        <>
-            <FormControlLabel
-                control={
-                    <Switch checked={dense} onChange={handleChangeDense} />
-                }
-                label="Dense padding"
-            />
-            <TableComponent
-                data={businessesList}
-                dataType="Business"
-                dense={dense}
-                headCells={headCells}
-            />
-        </>
+        <TableComponent
+            data={businessesList}
+            dataType="Business"
+            dense={dense}
+            headCells={headCells}
+            selected={selected}
+            setSelected={setSelected}
+        />
     );
 };
 
-const mapStateToProps = state => {
-    return {
-        businesses: state.businesses.businesses,
-        isLoggedIn: state.user.isLoggedIn,
-    };
-};
-
-export default connect(mapStateToProps, {
-    getBusinessesAct,
-})(BusinessesList);
+export default BusinessesList;
